requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1
torchvision==0.20.1
numpy==2.2.4
hydra-core==1.3.2
pytorch-lightning==2.4.0
rich==13.9.4
psutil==6.1.0
matplotlib==3.10.1
python-dotenv==1.0.1
scipy==1.14.1
scikit-learn==1.5.2
pytorch-msssim==1.0.0
opencv-python==4.10.0.84
albumentations==2.0.5
rasterio==1.4.2
scikit-image==0.24.0
geopandas==1.0.1
tensorboard==2.18.0
pycocotools==2.0.8
ipykernel==6.29.5